Fake DNS responses for some soup.io domains.
About half a day ago, soup.io began resolving to some spam site. But, it was discovered that the original server was still up. This commit has wget-lua fake a DNS reponse for the domain covered by the current item, and for asset.soup.io, in order to get those, and let this project continue.

This tries to be very conservative. The only "interesting" element of this is that it changes which of the two assets. servers it uses depending on the parity of the item name.